L1 - Purgatorio V1
Version one of the first level with some basic environmental art and collision, as well as a basic room for the boss of the level towards the end. Size/spacing of platforms can be adjusted as needed. Need help implementing door sprites that will lead to the next room in numerical order upon collision by player. All changed files will be in Raphael 4.27 because of the way my PC had me rebuild.